# Documentation/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Personal and family finance management system with multi-currency support and automatic balance calculation",
        "title": "Expense Tracker API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Igor Kudinov",
            "email": "digitlock@proton.me"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/currencies/convert": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Convert an amount from one currency to another using current exchange rates. Supports conversion between RSD and EUR.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Currencies"
                ],
                "summary": "Convert currency",
                "parameters": [
                    {
                        "type": "number",
                        "example": 100,
                        "description": "Amount to convert",
                        "name": "amount",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "RSD",
                            "EUR"
                        ],
                        "type": "string",
                        "example": "EUR",
                        "description": "Source currency code",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "RSD",
                            "EUR"
                        ],
                        "type": "string",
                        "example": "RSD",
                        "description": "Target currency code",
                        "name": "to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conversion result",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ConvertCurrencyResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid parameters (missing amount, invalid currency codes)",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing JWT token",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/currencies/rates": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve current exchange rates for all supported currencies (RSD, EUR) with base currency RSD",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Currencies"
                ],
                "summary": "Get exchange rates",
                "responses": {
                    "200": {
                        "description": "Current exchange rates",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ExchangeRatesResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing JWT token",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/exchange-rates/sync": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Triggers an immediate sync of exchange rates from the currency rate service and upserts them.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Currencies"
                ],
                "summary": "Force exchange rate sync",
                "responses": {
                    "200": {
                        "description": "Sync result",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ExchangeRateSyncResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing JWT token",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Currency rate service unavailable",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ConvertCurrencyResponse": {
            "type": "object",
            "properties": {
                "conversion_date": {
                    "type": "string",
                    "example": "2025-12-06T18:30:00Z"
                },
                "converted_amount": {
                    "type": "number",
                    "example": 11750
                },
                "exchange_rate": {
                    "type": "number",
                    "example": 117.5
                },
                "original_amount": {
                    "type": "number",
                    "example": 100
                },
                "original_currency": {
                    "type": "string",
                    "example": "EUR"
                },
                "target_currency": {
                    "type": "string",
                    "example": "RSD"
                }
            }
        },
...
        "dto.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ValidationError"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Invalid input"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/dto.ErrorDetail"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "dto.ExchangeRateSyncResponse": {
            "type": "object",
            "properties": {
                "fetched_at": {
                    "type": "string",
                    "example": "2025-12-06T18:30:00Z"
                },
                "source": {
                    "type": "string",
                    "example": "fawazahmed0"
                },
                "synced_pairs": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "dto.ExchangeRatesResponse": {
            "type": "object",
            "properties": {
                "base_currency": {
                    "type": "string",
                    "example": "RSD"
                },
                "last_updated": {
                    "type": "string",
                    "example": "2025-12-06T18:30:00Z"
                },
                "rates": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "number"
                    },
                    "example": {
                        "EUR": 0.008511,
                        "RSD": 1
                    }
                },
                "source": {
                    "type": "string",
                    "example": "api"
                }
            }
        },
...
        "dto.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "dto.ValidationError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string",
                    "example": "email"
                },
                "message": {
                    "type": "string",
                    "example": "Invalid email format"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}